Removed dependency on Boost program options and fixed remaining renaming issues
