+++ .gitsubprojects
@@ -1,5 +1,5 @@
 # -*- mode: cmake -*-
-git_subproject(Lunchbox https://github.com/Eyescale/Lunchbox.git 6bfee14)
-git_subproject(zeq https://github.com/HBPVis/zeq.git 36955ad)
+git_subproject(Lunchbox https://github.com/Eyescale/Lunchbox.git 66c6580)
+git_subproject(zeq https://github.com/HBPVis/zeq.git 17e71d4)
 git_subproject(FlatBuffers https://github.com/Bluebrain/flatbuffers.git b5721cf)
 git_subproject(cppnetlib https://github.com/cpp-netlib/cpp-netlib.git 0.11-devel)
