--- RESTBridge.config
@@ -1,1 +1,0 @@
-// ADD PREDEFINED MACROS HERE!
