--- RESTBridge.creator
@@ -1,1 +1,0 @@
-[General]
