--- RESTBridge.includes
@@ -1,2 +1,0 @@
-restconnector
-restconnector/detail
