+++ apps/CMakeLists.txt
@@ -5,6 +5,6 @@
 include_directories(${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_BINARY_DIR})
 
 set(RESTBRIDGE_SOURCES RESTBridge.cpp)
-set(RESTBRIDGE_LINK_LIBRARIES RESTBridge)
+set(RESTBRIDGE_LINK_LIBRARIES ${Boost_PROGRAM_OPTIONS_LIBRARY} RESTBridge)
 
 common_application(restBridge)
