+++ restbridge/CMakeLists.txt
@@ -17,7 +17,9 @@ set(RESTBRIDGE_SOURCES
   detail/RequestHandler.cpp
   detail/RestZeqTranslator.cpp)
 
-set(RESTBRIDGE_LINK_LIBRARIES ${Boost_LIBRARIES} ${LUNCHBOX_LIBRARIES}
+set(MINIMAL_BOOST_LIBRARIES ${Boost_LIBRARIES})
+list(REMOVE_ITEM MINIMAL_BOOST_LIBRARIES ${Boost_PROGRAM_OPTIONS_LIBRARY})
+set(RESTBRIDGE_LINK_LIBRARIES ${MINIMAL_BOOST_LIBRARIES} ${LUNCHBOX_LIBRARIES}
   ${ZEQ_LIBRARIES} ${CPPNETLIB_LIBRARIES} pthread)
 set(RESTBRIDGE_OMIT_PROJECT_HEADER ON)
