+++ restbridge/types.h
@@ -6,10 +6,10 @@
  * @remarks Copyright (c) BBP/EPFL 2005-2014; All rights reserved. Do not distribute without further notice.
  */
 
-#ifndef RESTCONNECTOR_TYPES_H
-#define RESTCONNECTOR_TYPES_H
+#ifndef RESTRESTBRIDGE_TYPES_H
+#define RESTRESTBRIDGE_TYPES_H
 
-namespace restconnector
+namespace restbridge
 {
 
 class RequestHandler;
@@ -17,4 +17,4 @@ class RestZeqTranslator;
 
 }
 
-#endif // RESTCONNECTOR_TYPES_H
+#endif // RESTBRIDGE_TYPES_H
